refactor: meta api, hooks 분리
