+++ src/services/meta/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './useCampuses';
+export * from './useTracks';
+export * from './useYears';
