+++ src/services/meta/hooks/useCampuses.ts
@@ -1,7 +1,5 @@
 import { useMemo } from 'react';
 
-import { SsafyTrack } from '~/services/member';
-
 export const initialCampuses = [
   { id: 1, name: '서울' },
   { id: 2, name: '대전' },
@@ -18,22 +16,4 @@ export const useCampuses = () => {
   }, []);
 
   return { data: campuses };
-};
-
-export const useTracks = () => {
-  const tracks = useMemo(() => Object.values(SsafyTrack), []);
-  return { data: tracks };
-};
-
-export const useYears = () => {
-  const maxYear = 10;
-  const years = useMemo(
-    () =>
-      Array(maxYear)
-        .fill(undefined)
-        .map((_, i) => i + 1),
-    []
-  );
-
-  return { data: years };
 };
