+++ src/services/meta/hooks/useTracks.ts
@@ -1,0 +1,8 @@
+import { useMemo } from 'react';
+
+import { SsafyTrack } from '~/services/member/utils';
+
+export const useTracks = () => {
+  const tracks = useMemo(() => Object.values(SsafyTrack), []);
+  return { data: tracks };
+};
